Create Exercise2
Power of 3
It enters a while loop that continues as long as '$x' is divisible by 3. Inside the loop, it divides '$x' by 3 in each iteration.
After the loop, it checks if '$x' is equal to 1. If it is, then '$n' is a power of 3, so it returns a string indicating that. Otherwise, it returns a string indicating that '$n' is not a power of 3.